wip - code structure refactoring
